[fix] Remove "upload image" button and modify pop up in layers configuration component
